+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+-  Adapt comparative category widget design [#575](https://github.com/CartoDB/carto-react/pull/575)
 - Improved Comparative Formula UI props interface [#573](https://github.com/CartoDB/carto-react/pull/573)
 
 ## 1.5